+++ app/routes/components/Popover/Demo.tsx
@@ -1,7 +1,6 @@
 import { useState } from "react";
 
-import { Button } from "src/components/Button";
-import { Popover, type Placement } from "src/components/Popover";
+import { type Placement, Popover } from "src/components/Popover";
 import { Select } from "src/components/Select";
 
 import styles from "./Demo.module.scss";
@@ -9,37 +8,39 @@ import styles from "./Demo.module.scss";
 export function Demo() {
   const [placement, setPlacement] = useState(null as Placement | null);
 
-  const placementOptions: Array<{ label: string, value: Placement }> = [
-    { label: 'Top Start', value: 'top-start' },
-    { label: 'Top', value: 'top' },
-    { label: 'Top End', value: 'top-end' },
-    { label: 'Right Start', value: 'right-start' },
-    { label: 'Right', value: 'right' },
-    { label: 'Right End', value: 'right-end' },
-    { label: 'Bottom Start', value: 'bottom-start' },
-    { label: 'Bottom', value: 'bottom' },
-    { label: 'Bottom End', value: 'bottom-end' },
-    { label: 'Left Start', value: 'left-start' },
-    { label: 'Left', value: 'left' },
-    { label: 'Left End', value: 'left-end' },
+  const placementOptions: Array<{ label: string; value: Placement }> = [
+    { label: "Top Start", value: "top-start" },
+    { label: "Top", value: "top" },
+    { label: "Top End", value: "top-end" },
+    { label: "Right Start", value: "right-start" },
+    { label: "Right", value: "right" },
+    { label: "Right End", value: "right-end" },
+    { label: "Bottom Start", value: "bottom-start" },
+    { label: "Bottom", value: "bottom" },
+    { label: "Bottom End", value: "bottom-end" },
+    { label: "Left Start", value: "left-start" },
+    { label: "Left", value: "left" },
+    { label: "Left End", value: "left-end" },
   ];
 
   return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: '5rem' }}>
+    <div style={{ display: "flex", flexDirection: "column", gap: "5rem" }}>
       <Popover placement={placement || undefined}>
         <Popover.Trigger
           render={({ setRef, onToggle }) => (
-            <div
-              ref={(el) => setRef(el)} onClick={onToggle}
+            <button
+              type="button"
+              ref={(el) => setRef(el)}
+              onClick={onToggle}
               className={styles.popoverTrigger}
             >
               Open Popover Portal
-            </div>
+            </button>
           )}
         />
         <Popover.Portal
           onClickOutside={({ closePortal: close }) => close()}
-          render={({ setRef, isOpen, floatingStyles }) => (
+          render={({ setRef, isOpen, floatingStyles }) =>
             isOpen && (
               <div
                 ref={(el) => setRef(el)}
@@ -49,7 +50,7 @@ export function Demo() {
                 You click to open me!
               </div>
             )
-          )}
+          }
         />
       </Popover>
       <Select