+++ app/routes/components/Popover/changelog.codegen.ts
@@ -12,4 +12,5 @@ export const changelog = ''
   + '  portal.\n'
   + "- 2025-09-25: Support `onOpen` and `onClose` props for `Popover.Trigger`'s\n"
   + '  render. And rename its `onClick` props to `onToggle`.\n'
+  + '- 2025-10-03: Fix missing dependencies in useEffect hooks.\n'
   ;