+++ src/components/Popover/CHANGELOG.md
@@ -8,3 +8,4 @@
   portal.
 - 2025-09-25: Support `onOpen` and `onClose` props for `Popover.Trigger`'s
   render. And rename its `onClick` props to `onToggle`.
+- 2025-10-03: Fix missing dependencies in useEffect hooks.